working.

+++ net/net-gossip.sh
@@ -226,6 +226,7 @@ CARGO_BIN="\$HOME/.cargo/bin"
 
 startCommon() {
   declare ipAddress=$1
+  declare instanceIndex=$2
   test -d "$SOLANA_ROOT"
   if $skipSetup; then
     # shellcheck disable=SC2029
@@ -238,11 +239,16 @@ startCommon() {
       mkdir -p $SOLANA_HOME $CARGO_BIN;
       mv ~/config $SOLANA_HOME/
     "
+  elif ! $instanceIndex; then
+      ssh "${sshOptions[@]}" "$ipAddress" "
+      set -x;
+      rm -rf $SOLANA_HOME;
+      mkdir -p $CARGO_BIN
+    "
   else
     # shellcheck disable=SC2029
     ssh "${sshOptions[@]}" "$ipAddress" "
       set -x;
-      rm -rf $SOLANA_HOME;
       mkdir -p $CARGO_BIN
     "
   fi
@@ -381,6 +387,9 @@ startGossipNode() {
   declare ipAddress=$1
   declare nodeType=$2
   declare nodeIndex="$3"
+  declare instanceIndex=$4
+
+  echo "startgossipnode() instance index: $instanceIndex"
 
   initLogDir
   declare logFile="$netLogDir/validator-$ipAddress.log"
@@ -399,7 +408,7 @@ startGossipNode() {
   echo "start log: $logFile"
   (
     set -x
-    startCommon "$ipAddress"
+    startCommon "$ipAddress" $instanceIndex
 
     if [[ $nodeType = blockstreamer ]] && [[ -n $letsEncryptDomainName ]]; then
       #
@@ -440,6 +449,7 @@ startGossipNode() {
          \"$waitForGossipNodeInit\" \
          \"$extraPrimordialStakes\" \
          \"$TMPFS_ACCOUNTS\" \
+         \"$instanceIndex\" \
       "
   ) >> "$logFile" 2>&1 &
   declare pid=$!
@@ -702,6 +712,8 @@ prepareDeploy() {
 
 gossipDeploy() {
   echo "greg - in gossipDeploy()"
+  declare instancesPerNode=$1
+  echo "greg - instancesPerNode in gossipDeploy: $instancesPerNode"
   initLogDir
 
   echo "Deployment started at $(date)"
@@ -723,12 +735,17 @@ gossipDeploy() {
       SECONDS=0
       pids=()
     else
-      startGossipNode "$ipAddress" $nodeType $nodeIndex
+      for (( i = 0; i < $instancesPerNode; i++ ))
+      do
+        startGossipNode "$ipAddress" $nodeType $nodeIndex $i
+        sleep 2
+      done
+      # startGossipNode "$ipAddress" $nodeType $nodeIndex
 
-      # Stagger additional node start time. If too many nodes start simultaneously
-      # the bootstrap node gets more rsync requests from the additional nodes than
-      # it can handle.
-      sleep 2
+      # # Stagger additional node start time. If too many nodes start simultaneously
+      # # the bootstrap node gets more rsync requests from the additional nodes than
+      # # it can handle.
+      # sleep 2
     fi
   done
 
@@ -1001,6 +1018,7 @@ maybeFullRpc=false
 waitForNodeInit=true
 waitForGossipNodeInit=false
 extraPrimordialStakes=0
+instancesPerNode=1 # default. 1 replica per node
 
 command=$1
 [[ -n $command ]] || usage
@@ -1130,6 +1148,9 @@ while [[ -n $1 ]]; do
     elif [[ $1 = --skip-require-tower ]]; then
       maybeSkipRequireTower="$1"
       shift 1
+    elif [[ $1 = --gossip-instances-per-node ]]; then
+      instancesPerNode="$2"
+      shift 2
     else
       usage "Unknown long option: $1"
     fi
@@ -1138,6 +1159,8 @@ while [[ -n $1 ]]; do
     shift
   fi
 done
+
+echo "greg - gossip instances per node: $instancesPerNode"
 
 while getopts "h?T:t:o:f:rc:Fn:i:d" opt "${shortArgs[@]}"; do
   case $opt in
@@ -1276,7 +1299,7 @@ case $command in
 gossip-only)
   prepareDeploy
   stop
-  gossipDeploy
+  gossipDeploy $instancesPerNode
   ;;
 restart)
   prepareDeploy
